requirements.txt
# Meetings2Obsidian Dependencies
# Python 3.12+ required (REQUIREMENTS.md specifies 3.14, but using 3.12 as 3.14 is not yet released)

# Configuration management
pyyaml>=6.0.1

# HTTP requests for API calls (Heypocket)
requests>=2.31.0

# Browser automation for Google Meet and Zoom
playwright>=1.40.0

# Note: After installing dependencies, run:
# playwright install chromium
# This will download the Chromium browser needed for automation
